backend/requirements.txt
flask==3.0.2
requests==2.32.3
python-dotenv==1.0.1
pinecone-client==4.1.0
